Move EGRID test helper into rft_generator
Lift the parameterized egrid content builder into the shared
`tests/ert/rft_generator` module as `create_egrid`, removing duplication
and giving future RFT tests a common function for creating synthetic
EGRIDs.
